requirements_local.txt
flask
flask-cors
psycopg2-binary
numpy
pandas
fpdf
openpyxl
gunicorn
google-genai
python-dotenv
opencv-python
pyttsx3
mediapipe
requests